requirements.txt
asgiref==3.5.1
boto==2.49.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
cloudinary==1.29.0
dj-static==0.0.6
dj3-cloudinary-storage==0.0.6
Django==3.0
django-chunked-upload==2.0.0
django-crispy-forms==1.14.0
django-markdown==0.8.4
django-pagedown==2.2.1
django-storages==1.12.3
django-tinymce==3.4.0
google-api-core==2.7.3
google-auth==2.6.6
google-cloud-core==2.3.0
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
Markdown==3.3.6
Pillow==9.1.0
protobuf==3.20.1
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-decouple==3.6
python-dotenv==0.20.0
python-magic==0.4.25
pytz==2022.1
requests==2.27.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
static3==0.7.0
urllib3==1.26.9
xlwt==1.3.0
zipp==3.8.0
